refactor bootstrap
