+++ src/ui/app.js
@@ -1,3 +1,25 @@
-// Core UI bootstrap placeholder.
-// Exported for validation purposes; real implementation pending.
-export function initApp() {}
+import { createGameController } from "../game/GameController.js";
+import { mountAllFeatureUIs } from "../features/index.js";
+import { mountDevQuickMenu } from "./dev/devQuickMenu.js";
+
+// Bootstraps the game controller, mounts feature UIs and starts the loop.
+export function initApp() {
+  const game = createGameController();
+  mountAllFeatureUIs(game.state);
+  game.start();
+
+  // window.game = game; // optional for debugging
+
+  window.__PAUSE = () => game.pause();
+  window.__RESUME = () => game.resume();
+  window.__STEP = () => game.step();
+  window.__GET_SPEED = () => game.getSpeed();
+  window.__SET_SPEED = (v) => game.setSpeed(v);
+
+  if (document.readyState === "loading") {
+    document.addEventListener("DOMContentLoaded", () => mountDevQuickMenu());
+  } else {
+    mountDevQuickMenu();
+  }
+}
+
